feat: add make file and split to multifile
